+++ geotrek/infrastructure/templates/infrastructure/sql/post_90_defaults.sql
@@ -38,3 +38,4 @@ ALTER TABLE infrastructure_infrastructure ALTER COLUMN description SET DEFAULT '
 ALTER TABLE infrastructure_infrastructure ALTER COLUMN published SET DEFAULT FALSE;
 -- publication_date
 -- structure
+ALTER TABLE infrastructure_infrastructure ALTER COLUMN provider SET DEFAULT '';